apply consisten folder naming
