+++ tests/end-to-end/external-solution-integration/internal/scenario/compass/compass.go
@@ -1,4 +1,4 @@
-package compass_e2e
+package compass
 
 import (
 	"k8s.io/client-go/rest"
