+++ tests/end-to-end/external-solution-integration/internal/scenario/compass/scenario.go
@@ -1,4 +1,4 @@
-package connectivity_adapter_e2e
+package compass
 
 import (
 	"github.com/kyma-project/kyma/tests/end-to-end/external-solution-integration/internal/scenario"
@@ -7,6 +7,8 @@ import (
 	"github.com/vrischmann/envconfig"
 )
 
+// Scenario executes complete external solution integration test scenario
+// using Compass for Application registration and connectivity
 type Scenario struct {
 	domain        string
 	testID        string
@@ -16,7 +18,7 @@ type Scenario struct {
 // AddFlags adds CLI flags to given FlagSet
 func (s *Scenario) AddFlags(set *pflag.FlagSet) {
 	pflag.StringVar(&s.domain, "domain", "kyma.local", "domain")
-	pflag.StringVar(&s.testID, "testID", "connectivity-adapter-e2e", "domain")
+	pflag.StringVar(&s.testID, "testID", "compass-e2e-test", "domain")
 	pflag.BoolVar(&s.skipSSLVerify, "skipSSLVerify", false, "Skip verification of service SSL certificates")
 }
 
@@ -26,6 +28,7 @@ func (s *Scenario) NewState() (*state, error) {
 	if err != nil {
 		return nil, errors.Wrap(err, "while loading environment variables")
 	}
+
 	return &state{
 		E2EState:         scenario.E2EState{Domain: s.domain, SkipSSLVerify: s.skipSSLVerify, AppName: s.testID},
 		CompassEnvConfig: config,
