+++ tests/end-to-end/external-solution-integration/internal/scenario/connectivity_adapter/state.go
@@ -1,13 +1,14 @@
-package compass_e2e
+package connectivity_adapter
 
 import (
 	"crypto/tls"
 	"fmt"
 	"net/http"
 
+	"github.com/kyma-project/kyma/tests/end-to-end/external-solution-integration/internal/scenario"
+
 	"github.com/kyma-project/kyma/common/resilient"
 	"github.com/kyma-project/kyma/tests/end-to-end/external-solution-integration/internal"
-	"github.com/kyma-project/kyma/tests/end-to-end/external-solution-integration/internal/scenario"
 	"github.com/kyma-project/kyma/tests/end-to-end/external-solution-integration/pkg/testkit"
 )
 
@@ -15,16 +16,27 @@ type state struct {
 	scenario.E2EState
 	scenario.CompassEnvConfig
 	compassAppID         string
-	connectivityURL      string
 	servicePlanID        string
 	registryClient       *testkit.RegistryClient
 	legacyRegistryClient *testkit.LegacyRegistryClient
 	cert                 []tls.Certificate
 }
 
-// GetCompassAppID returns Compass ID of registered application
-func (s *state) GetCompassAppID() string {
-	return s.compassAppID
+func (s *state) GetRegistryClient() *testkit.RegistryClient {
+	return s.registryClient
+}
+
+func (s *state) GetEventSender() *testkit.EventSender {
+	return s.EventSender
+}
+
+func (s *state) SetGatewayClientCerts(certs []tls.Certificate) {
+	httpClient := internal.NewHTTPClient(s.SkipSSLVerify)
+	httpClient.Transport.(*http.Transport).TLSClientConfig.Certificates = certs
+	resilientHTTPClient := resilient.WrapHttpClient(httpClient)
+	legacyMetadataURL := fmt.Sprintf("https://adapter-gateway-mtls.%s/%s/v1/metadata/services", s.Domain, s.AppName)
+	s.registryClient = testkit.NewRegistryClient(legacyMetadataURL, resilientHTTPClient)
+	s.EventSender = testkit.NewEventSender(resilientHTTPClient, s.Domain, nil)
 }
 
 // SetCompassAppID sets Compass ID of registered application
@@ -32,13 +44,9 @@ func (s *state) SetCompassAppID(appID string) {
 	s.compassAppID = appID
 }
 
-func (s *state) SetGatewayClientCerts(certs []tls.Certificate) {
-	metadataURL := fmt.Sprintf("https://gateway.%s/%s/v1/metadata/services", s.Domain, s.AppName)
-	httpClient := internal.NewHTTPClient(s.SkipSSLVerify)
-	httpClient.Transport.(*http.Transport).TLSClientConfig.Certificates = certs
-	resilientHTTPClient := resilient.WrapHttpClient(httpClient)
-	s.registryClient = testkit.NewRegistryClient(metadataURL, resilientHTTPClient)
-	s.EventSender = testkit.NewEventSender(resilientHTTPClient, s.Domain, nil)
+// GetCompassAppID returns Compass ID of registered application
+func (s *state) GetCompassAppID() string {
+	return s.compassAppID
 }
 
 func (s *state) GetServicePlanID() string {
