+++ tests/end-to-end/external-solution-integration/internal/scenario/event_mesh/scenario.go
@@ -1,4 +1,4 @@
-package event_mesh_e2e
+package event_mesh
 
 import (
 	"github.com/kyma-project/kyma/tests/end-to-end/external-solution-integration/internal/scenario"
